Refactor utils to resolve Node dependency issue
