+++ src/test/pages/SetAssignments.test.tsx
@@ -2,7 +2,9 @@ import {screen, waitFor, within} from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import {SetAssignments} from "../../app/components/pages/SetAssignments";
 import {mockActiveGroups, mockGameboards, mockSetAssignments} from "../../mocks/data";
-import {dayMonthYearStringToDate, DDMMYYYY_REGEX, ONE_DAY_IN_MS, renderTestEnvironment} from "../utils";
+import {dayMonthYearStringToDate, DDMMYYYY_REGEX, ONE_DAY_IN_MS} from "../dateUtils";
+import {renderTestEnvironment} from "../testUtils";
+
 import {API_PATH, isAda, isPhy, PATHS, siteSpecific} from "../../app/services";
 import {rest} from "msw";
